# api/swagger.yaml
basePath: /v1
definitions:
  domain.Account:
    properties:
      balance_cents:
        description: Always in cents; never float
        type: integer
      created_at:
        type: string
      currency:
        description: ISO 4217 code (e.g. "USD")
        type: string
      deleted_at:
        type: string
      id:
        type: string
      name:
        type: string
      tenant_id:
        type: string
      type:
        $ref: '#/definitions/domain.AccountType'
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  domain.AccountType:
    enum:
    - checking
    - savings
    - credit_card
    - investment
    type: string
    x-enum-varnames:
    - AccountTypeChecking
    - AccountTypeSavings
    - AccountTypeCreditCard
    - AccountTypeInvestment
...
  handler.CreateAccountRequest:
    properties:
      currency:
        type: string
      initial_cents:
        type: integer
      name:
        maxLength: 100
        minLength: 1
        type: string
      type:
        allOf:
        - $ref: '#/definitions/domain.AccountType'
        enum:
        - checking
        - savings
        - credit_card
        - investment
    required:
    - currency
    - initial_cents
    - name
    - type
    type: object
...
  handler.UpdateAccountRequest:
    properties:
      currency:
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: API Support
  description: Moolah Household Finance & Investment SaaS API.
  title: Moolah Financial API
  version: "1.0"
paths:
  /v1/accounts:
    get:
      consumes:
      - application/json
      description: Returns all accounts belonging to the current tenant.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Account'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List accounts
      tags:
      - accounts
    post:
      consumes:
      - application/json
      description: Creates a new financial account for the current tenant.
      parameters:
      - description: Optional idempotency key (ULID format recommended)
        in: header
        name: Idempotency-Key
        type: string
      - description: Account details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Account'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create account
      tags:
      - accounts
  /v1/accounts/{id}:
    delete:
      consumes:
      - application/json
      description: Soft-deletes an account and its transactions.
      parameters:
      - description: Account ULID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Account deleted
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Account not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete account
      tags:
      - accounts
    get:
      consumes:
      - application/json
      description: Returns details of a specific account by ID.
      parameters:
      - description: Account ULID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Account'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Account not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get account
      tags:
      - accounts
    patch:
      consumes:
      - application/json
      description: Updates details of an existing account.
      parameters:
      - description: Account ULID
        in: path
        name: id
        required: true
        type: string
      - description: Optional idempotency key (ULID format recommended)
        in: header
        name: Idempotency-Key
        type: string
      - description: Update fields
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Account'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Account not found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update account
      tags:
      - accounts
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"